+++ plugin/McpLink/Source/McpLinkEditor/McpLinkEditor.Build.cs
@@ -22,6 +22,8 @@ public class McpLinkEditor : ModuleRules
 			"Json",
 			"JsonUtilities",
 			"McpLinkCore",
+			"AutomationDriver",
+			"InputCore",
 			"UnrealEd",
 			"AssetRegistry",
 			"AssetTools",